Enable password grant type
When someone begins editing a Sandbox project (for example via Tasking
Manager), their client contacts the Dashboard to authenticate. The
Dashboard will create a user in the given Sandbox whose username matches
the person's OSM username, and set the password == that username.

Afterwards, the Dashboard needs to be able to _use_ those credentials
to retrieve an OAuth token for the sandbox, and return it back to the
client. For this to work, the 'password' grant type must be enabled. The
OpenStreetMap Rails app disables this for security reasons, so we need
to patch the code to re-enable it.